+++ lua/plugins/debug-vimspector.lua
@@ -1,29 +1,8 @@
 local utils = require("utils")
----core code
-local start
-local quit
-
-local normalKeys = {
-  {
-    "<leader>dss",
-    function()
-      start()
-    end,
-    desc = "Start vim-spector debugging",
-  },
-  { "<leader>dsb", "<Plug>VimspectorToggleBreakpoint", desc = "Toggle line breakpoint on the current line" },
-  { "<leader>dsf", "<Plug>VimspectorAddFunctionBreakpoint", desc = "Add a function breakpoint" },
-}
 
 local debuggingKeys = {
-  {
-    "<leader><F5>",
-    function()
-      quit()
-    end,
-    desc = "Quit vim-spector debugging",
-  },
-  { "<F5>", "<Plug>VimspectorContinue", desc = "Continue" },
+  { "<leader><F5>", "<cmd>VimspectorReset<cr>", desc = "Quit vim-spector debugging" },
+  { "<F5>", "<Plug>VimspectorContinue", desc = "Start or Continue" },
   { "<F17>", "<Plug>VimspectorStop", desc = "<S-F5> Stop debugging" },
   { "<F41>", "<Plug>VimspectorRestart", desc = "<C-S-F5> Restart debugging with the same configuration" },
   { "<F8>", "<Plug>VimspectorJumpToNextBreakpoint", desc = "Jump to next breakpoint" },
@@ -40,23 +19,35 @@ local debuggingKeys = {
   { "<leader>di", "<Plug>VimspectorBalloonEval", desc = "Debug Inspect", mode = { "n", "x" } },
 }
 
-start = function()
-  vim.cmd("call vimspector#Continue()")
-  utils.deleteKeys(normalKeys)
-  utils.mappingKeys(debuggingKeys)
-end
-
-quit = function()
-  vim.cmd("call vimspector#Reset()")
-  utils.deleteKeys(debuggingKeys)
-  utils.mappingKeys(normalKeys)
-end
-
 vim.cmd("let g:vimspector_base_dir='/home/jover/.local/share/nvim/lazy/vimspector'")
 
 return {
   "puremourning/vimspector",
+  init = function()
+    require("which-key").register({
+      ["<leader>ds"] = { name = "vimspector" },
+    })
+  end,
   keys = function()
-    return normalKeys
+    local running = false
+    return {
+      {
+        "<leader>dss",
+        function()
+          if running then
+            -- stop
+            utils.deleteKeys(debuggingKeys)
+            running = false
+            print("=== debugger vim-spector is Stopped ===")
+          else
+            -- start
+            utils.mappingKeys(debuggingKeys)
+            running = true
+            print("=== debugger vim-spector is Started ===")
+          end
+        end,
+        desc = "Start or Stop vim-spector debugging",
+      },
+    }
   end,
 }
